+++ com.unity.formats.alembic/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed bug where vertex color displayed as black when processing RGB, RGBA, and Normal attributes in the `Face` scope
 - Fixed bug causing Editor to crash when a UV attribute is processed in the `Points` or `Face` scope.
+- Prevent a NullReferenceException when adding a curve renderer to an Alembic file with non-visible meshes.
 
 ## [2.3.4] - 2023-10-30
 ### Fixed